feat: crud products

+++ services/products/controllers/controller.js
@@ -1,17 +1,18 @@
 const { Product } = require("../models/index");
 
-class ProductController {
+class Controller {
   static async getProduct(req, res, next) {
     try {
       const dataProduct = await Product.findAll();
       res.status(200).json(dataProduct);
-    } catch (error) {
-      res.status(500).json(error);
+    } catch (err) {
+      next(err);
     }
   }
   static async postProduct(req, res, next) {
     try {
-      const { name, price, description, imgUrl, stock, CategoryId } = req.body;
+      const { name, price, description, imgUrl, stock, CategoryId, color } =
+        req.body;
       const product = await Product.create({
         name,
         price,
@@ -19,19 +20,43 @@ class ProductController {
         imgUrl,
         stock,
         CategoryId,
+        color,
       });
-      res.status(201).json({
-        message: `Product with name ${product.name} has been created`,
-      });
-    } catch (error) {
-      res.status(500).json(error);
+
+      res.status(201).json(product);
+    } catch (err) {
+      next(err);
+    }
+  }
+  static async getDetailProduct(req, res, next) {
+    try {
+      const { id } = req.params;
+      const productById = await Product.findByPk(id);
+
+      if (!productById) {
+        throw {
+          name: "Product Not Found",
+        };
+      }
+
+      res.status(200).json(productById);
+    } catch (err) {
+      next(err);
     }
   }
   static async updateProduct(req, res, next) {
     try {
       const { id } = req.params;
-      const { name, price, description, imgUrl, stock, CategoryId } = req.body;
+      const { name, price, description, imgUrl, stock, CategoryId, color } =
+        req.body;
       const productById = await Product.findByPk(id);
+
+      if (!productById) {
+        throw {
+          name: "Product Not Found",
+        };
+      }
+
       await Product.update(
         {
           name,
@@ -40,6 +65,7 @@ class ProductController {
           imgUrl,
           stock,
           CategoryId,
+          color,
         },
         {
           where: {
@@ -47,30 +73,38 @@ class ProductController {
           },
         }
       );
+
       res.status(200).json({
         message: `Product with name ${productById.name} has been updated`,
       });
-    } catch (error) {
-      res.status(500).json(error);
+    } catch (err) {
+      next(err);
     }
   }
   static async deleteProduct(req, res, next) {
     try {
       const { id } = req.params;
       const productById = await Product.findByPk(id);
+
+      if (!productById) {
+        throw {
+          name: "Product Not Found",
+        };
+      }
+
       await Product.destroy({
         where: {
           id,
         },
       });
+
       res.status(200).json({
         message: `Product with name ${productById.name} has been deleted`,
       });
-    } catch (error) {
-      console.log(error);
-      res.status(500).json(error);
+    } catch (err) {
+      next(err);
     }
   }
 }
 
-module.exports = ProductController;
+module.exports = Controller;
